requirements.txt
pandas==2.2.3
numpy==1.26.4
scipy==1.13.1
matplotlib==3.9.2
duckduckgo_search==6.3.4
aiohttp==3.11.12
beautifulsoup4==4.13.3
requests==2.32.3
python-dotenv==1.0.1
chromadb==0.5.18
llama-index==0.11.22
llama-parse==0.5.13
openai==1.61.1
aider-chat @ git+https://github.com/Aider-AI/aider.git@86175a18279c2fc9421998208d0e83f6677f6b47
litellm==1.60.8
groq==0.11.0
anthropic==0.28.1
llama-index-vector-stores-chroma==0.3.0
llama-index-llms-anthropic==0.2.1
llama-index-llms-azure-openai==0.2.2
llama-index-llms-bedrock==0.2.6
llama-index-llms-gemini==0.3.7
llama-index-llms-groq==0.2.0
llama-index-llms-openai==0.2.16
llama-index-embeddings-azure-openai==0.2.5
llama-index-embeddings-openai==0.2.5
httpx==0.27.2
# torch==2.5.1+cu121
# torchvision==0.20.1+cu121
# torchaudio==2.5.1+cu121
#marker-pdf==0.3.10
